+++ projects/fhi-angular-highcharts/src/lib/fhi-angular-highcharts.component.ts
@@ -32,6 +32,7 @@ import { OptionsService } from './services/options.service';
 import { TableService } from './services/table.service';
 import { DiagramTypeService } from './services/diagram-type.service';
 import { TopoJsonService } from './services/topo-json.service';
+import { TableData } from './models/table-data.model';
 
 @Component({
   selector: 'fhi-angular-highcharts',
@@ -56,8 +57,7 @@ export class FhiAngularHighchartsComponent implements OnChanges {
   currentDiagramTypeGroup!: string;
   diagramTypeGroups = DiagramTypeGroups;
   diagramTypeNavId = DiagramTypeNavIds;
-  tableHeaderRows = [];
-  tableBodyRows = [];
+  tableData: TableData;
 
   constructor(
     private changeDetector: ChangeDetectorRef,
@@ -229,8 +229,7 @@ export class FhiAngularHighchartsComponent implements OnChanges {
 
   private updateTable() {
     const series: FhiDiagramSerie[] = this.allDiagramOptions.series;
-    this.tableHeaderRows = this.tableService.getHeaderRows(series);
-    this.tableBodyRows = this.tableService.getDataRows(series);
+    this.tableData = this.tableService.getTable(series, this.allDiagramOptions.tableOrientation);
   }
 
   private updateMap() {