+++ src/MOCK_DB_DATA/library-items/accordions/accordion.ts
@@ -26,23 +26,20 @@ function getExampleHtml(): string {
  */
 function getCodeHtml(): string | null {
   return `
-<ngb-accordion>
-  <ngb-panel id="custom-id-0" title="Element #1">
-    <ng-template ngbPanelContent>
-      <p>Innhold for element #1</p>
-    </ng-template>
-  </ngb-panel>
-  <ngb-panel id="custom-id-1" title="Element #2">
-    <ng-template ngbPanelContent>
-      <p>Innhold for element #2</p>
-    </ng-template>
-  </ngb-panel>
-  <ngb-panel id="custom-id-2" title="Element #3">
-    <ng-template ngbPanelContent>
-      <p>Innhold for element #3</p>
-    </ng-template>
-  </ngb-panel>
-</ngb-accordion>`;
+<div ngbAccordion>
+  @for (item of accordionItems; track item) {
+    <div ngbAccordionItem [collapsed]="item !== 'First'">
+      <h2 ngbAccordionHeader>
+        <button ngbAccordionButton>Element #{{ item }}</button>
+      </h2>
+      <div ngbAccordionCollapse>
+        <div ngbAccordionBody>
+          <ng-template>Innhold element #{{ item }}</ng-template>
+        </div>
+      </div>
+    </div>
+  }
+</div>`;
 }
 
 /*