+++ src/MOCK_DB_DATA/library-items/alerts/alert-basic.ts
@@ -18,35 +18,35 @@ export const AlertBasic: LibraryItem[] = [
  */
 function getExampleHtml(): string {
   return `
-<div class="alert alert-info">
+<div class="alert alert-info" role="alert">
   <i class="icon-info-circle"></i>
   Dette er en melding til informasjon.
 </div>
-<div class="alert alert-success">
+<div class="alert alert-success" role="alert">
   <i class="icon-check-circle"></i>
   Dette er en melding om suksess!
 </div>
-<div class="alert alert-warning">
+<div class="alert alert-warning" role="alert">
   <i class="icon-bell"></i>
   Dette er en advarsel!
 </div>
-<div class="alert alert-error">
+<div class="alert alert-error" role="alert">
   <i class="icon-exclamation-circle"></i>
   Dette er en feilmelding.
 </div>
-<div class="alert alert-info fhi-alert-bordered">
+<div class="alert alert-info fhi-alert-bordered" role="alert">
   <i class="icon-info-circle"></i>
   Dette er en melding til informasjon.
 </div>
-<div class="alert alert-success fhi-alert-bordered">
+<div class="alert alert-success fhi-alert-bordered" role="alert">
   <i class="icon-check-circle"></i>
   Dette er en melding om suksess!
 </div>
-<div class="alert alert-warning fhi-alert-bordered">
+<div class="alert alert-warning fhi-alert-bordered" role="alert">
   <i class="icon-bell"></i>
   Dette er en advarsel!
 </div>
-<div class="alert alert-error fhi-alert-bordered">
+<div class="alert alert-error fhi-alert-bordered" role="alert">
   <i class="icon-exclamation-circle"></i>
   Dette er en feilmelding.
 </div>`;
@@ -58,13 +58,13 @@ function getExampleHtml(): string {
  */
 function getCodeHtml(): string | null {
   return `
-<div class="alert alert-info">
-  <i class="icon-..."></i> <!-- icon-info-circle | icon-check-circle | icon-bell | icon-exclamation-circle -->
+<div class="alert alert-info" role="alert">
+  <i class="icon-[name]"></i>
   Alert!
 </div>
 
-<div class="alert alert-info fhi-alert-bordered">
-  <i class="icon-..."></i>
+<div class="alert alert-info fhi-alert-bordered" role="alert">
+  <i class="icon-[name]"></i>
   Alert!
 </div>`;
 }